backend/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
dataclasses-json==0.6.7
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.9.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
intel-openmp==2021.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-google-genai==1.0.7
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langsmith==0.1.83
looseversion==1.3.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
Pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
prov==2.0.1
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.8.2
pydantic-core==2.20.1
pydeck==0.9.1
pydot==3.0.1
Pygments==2.18.0
PyMuPDF==1.24.7
pymupdfb==1.24.6
pyparsing==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.38.0
streamlit==1.36.0
sympy==1.12.1
tbb==2021.13.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traits==6.3.2
transformers==4.42.3
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.3
watchdog==4.0.1
Werkzeug==3.0.3
yarl==1.9.4